refactor : 함수명 네이밍 변경 및 BindingAdapter 위치 변경
